Split the library into modules
