+++ src/link.rs
@@ -1,0 +1,3 @@
+pub struct Link {
+    pub href: String,
+}
