+++ src/person.rs
@@ -1,0 +1,1 @@
+pub struct Person { }
